+++ kong/conf_loader/listeners.lua
@@ -15,8 +15,9 @@ local listeners = {}
 
 local subsystem_flags = {
   http = { "ssl", "http2", "proxy_protocol", "deferred", "bind", "reuseport",
-           "backlog=%d+" },
-  stream = { "udp", "ssl", "proxy_protocol", "bind", "reuseport", "backlog=%d+" },
+           "backlog=%d+", "ipv6only=on", "ipv6only=off" },
+  stream = { "udp", "ssl", "proxy_protocol", "bind", "reuseport", "backlog=%d+",
+             "ipv6only=on", "ipv6only=off" },
 }